+++ packages/lumx-react/src/components/progress-tracker/ProgressTrackerStepPanel.test.tsx
@@ -9,7 +9,8 @@ const CLASSNAME = ProgressTrackerStepPanel.className as string;
 
 // Mock useTabProviderContext.
 jest.mock('../tabs/state', () => {
-    return { useTabProviderContext: jest.fn(), useTabProviderContextState: jest.fn() };
+    const state = jest.requireActual('../tabs/state');
+    return { ...state, useTabProviderContext: jest.fn(), useTabProviderContextState: jest.fn() };
 });
 
 type SetupProps = Partial<ProgressTrackerStepPanelProps>;