+++ packages/lumx-react/src/components/tabs/TabList.test.tsx
@@ -14,7 +14,8 @@ const CLASSNAME = TabList.className as string;
 
 // Mock useTabProviderContext.
 jest.mock('./state', () => {
-    return { useTabProviderContext: jest.fn(), useTabProviderContextState: jest.fn() };
+    const state = jest.requireActual('./state');
+    return { ...state, useTabProviderContext: jest.fn(), useTabProviderContextState: jest.fn() };
 });
 
 type SetupProps = Partial<TabListProps>;